chore: move routes to a separate file
